rename routes/datasets to routes/list_datasets
