+++ src/routes/mod.rs
@@ -1,11 +1,11 @@
-mod datasets;
 mod gtfs_rt;
+mod list_datasets;
 mod status;
 mod stop_monitoring;
 mod stoppoints_discovery;
 
-pub use self::datasets::list_datasets;
 pub use self::gtfs_rt::{gtfs_rt, gtfs_rt_json};
+pub use self::list_datasets::list_datasets;
 pub use self::status::status_query;
 pub use self::stop_monitoring::stop_monitoring_query;
 pub use self::stoppoints_discovery::sp_discovery;
